feat: add codec name to codec code helper
Introduce the function `getCodeVarint()` to return a Buffer containing
the Varint encoded codec code of a given codec name. It throws an
error if the codec isn't found.

Reason for this change was the js-cid module that is using the
conversion table directly. If the codec doesn't exist it would throw
error messages that didn't make much sense.