chore: fully automatic table generation
The codec table is now created fully automatically. You just need
to run a single script. The script is written in Python in hope
that it can be adapted to be used for other multicodec implementations
as well. Python is most likely installed on all systems and no
big toolchain is needed in order to run/modify it.

This commit also updates the table to the most current version.

BREAKING CHANGE: multibase is not part of this package anymore

As multibase works differently from multicodec, those codecs
were [removed from the multicodec table], hence those are also
removed from this implementation as we have automatic
conversion from the upstream table.

[removed from the multicodec table]:
https://github.com/multiformats/multicodec/commit/1ec0e971d589d2fd5d5418b212846301909525bd